ui: move TraceStream to public API
If the public App API is to add a method to load a trace from a,
TraceStream then that TraceStream API likewise should be public.
